feat: dynamically retrieve initial data for app
Remove hardcoded testData and implement automatic retrieval of the
analysis for the previous seven days. Add analysis caching to reduce
the number of database queries and greatly improve the load time for
the initial data and any repeated queries. Refactor related code to
coordinate with this and to better style.